Move CircuitGraph.hpp to Model/PowerElectronics
